re-commit changes

+++ models/networks.py
@@ -283,7 +283,7 @@ class Encoder(nn.Module):
         outputs_mean = outputs.clone()        
         inst_list = np.unique(inst.cpu().numpy().astype(int))        
         for i in inst_list:
-            indices = (inst == i).nonzero() # n x 4            
+            indices = (inst == np.int(i)).nonzero() # n x 4            
             for j in range(self.output_nc):
                 output_ins = outputs[indices[:,0], indices[:,1] + j, indices[:,2], indices[:,3]]                    
                 mean_feat = torch.mean(output_ins).expand_as(output_ins)
